Move tools from base-preset to helper
