+++ package.json
@@ -8,8 +8,7 @@
   ],
   "scripts": {
     "commit": "git-cz",
-    "test:all": "cross-env NODE_ENV='test' lerna run build && nyc --reporter=json --reporter=text --reporter=lcov mocha packages/*/test presets/*/test && ./fix-coverage.sh",
-    "//test:all": "cross-env NODE_ENV='test' lerna run build && nyc --reporter=json mocha packages/*/test presets/*/test && remap-istanbul -i coverage/coverage-final.json -o coverage/coverage.json",
+    "test:all": "cross-env NODE_ENV=test lerna run build && nyc --reporter=json --reporter=text --reporter=lcov mocha packages/*/test presets/*/test && ./fix-coverage.sh",
     "coverage": "codecov && cat ./coverage/lcov.info | codacy-coverage"
   },
   "devDependencies": {
@@ -20,16 +19,11 @@
     "cz-conventional-changelog": "^2.1.0",
     "lerna": "^2.5.1",
     "mocha": "^5.0.0",
-    "nyc": "10.0.0",
-    "remap-istanbul": "^0.10.1"
+    "nyc": "10.0.0"
   },
   "config": {
     "commitizen": {
       "path": "node_modules/cz-conventional-changelog"
     }
-  },
-  "//nyc": {
-    "sourceMap": false,
-    "instrument": false
   }
 }
